Generate a table from the DESeq2 results

Generate a tidy table with the results of DESeq2

Usage

deseqresult2df(res_de, FDR = NULL)

Arguments

res_de

A DESeqResults object.

FDR

Numeric value, specifying the significance level for thresholding adjusted p-values. Defaults to NULL, which would return the full set of results without performing any subsetting based on FDR.

Value

A tidy data.frame with the results from differential expression, sorted by adjusted p-value. If FDR is specified, the table contains only genes with adjusted p-value smaller than the value.

Examples

library(airway)
library(DESeq2)
data(res_airway, package = "mosdef")
head(res_airway)
#> log2 fold change (MLE): dex untrt vs trt 
#> Wald test p-value: dex untrt vs trt 
#> DataFrame with 6 rows and 8 columns
#>                   baseMean log2FoldChange     lfcSE      stat      pvalue
#>                  <numeric>      <numeric> <numeric> <numeric>   <numeric>
#> ENSG00000000003 708.602170      0.3812540  0.100654  3.787752 0.000152016
#> ENSG00000000005   0.000000             NA        NA        NA          NA
#> ENSG00000000419 520.297901     -0.2068126  0.112219 -1.842943 0.065337292
#> ENSG00000000457 237.163037     -0.0379204  0.143445 -0.264356 0.791505742
#> ENSG00000000460  57.932633      0.0881682  0.287142  0.307054 0.758801924
#> ENSG00000000938   0.318098      1.3782270  3.499873  0.393793 0.693733530
#>                       padj      SYMBOL      symbol
#>                  <numeric> <character> <character>
#> ENSG00000000003 0.00128292      TSPAN6      TSPAN6
#> ENSG00000000005         NA        TNMD        TNMD
#> ENSG00000000419 0.19646985        DPM1        DPM1
#> ENSG00000000457 0.91141962       SCYL3       SCYL3
#> ENSG00000000460 0.89500478       FIRRM       FIRRM
#> ENSG00000000938         NA         FGR         FGR
res_df <- deseqresult2df(res_airway)
head(res_df)
#>                              id   baseMean log2FoldChange     lfcSE      stat
#> ENSG00000152583 ENSG00000152583   997.4398      -4.574919 0.1840561 -24.85611
#> ENSG00000165995 ENSG00000165995   495.0929      -3.291062 0.1331737 -24.71255
#> ENSG00000120129 ENSG00000120129  3409.0294      -2.947810 0.1214377 -24.27426
#> ENSG00000101347 ENSG00000101347 12703.3871      -3.766995 0.1554380 -24.23472
#> ENSG00000189221 ENSG00000189221  2341.7673      -3.353580 0.1417823 -23.65301
#> ENSG00000211445 ENSG00000211445 12285.6151      -3.730403 0.1658306 -22.49526
#>                        pvalue          padj  SYMBOL  symbol
#> ENSG00000152583 2.220933e-136 4.001677e-132 SPARCL1 SPARCL1
#> ENSG00000165995 7.839410e-135 7.062524e-131  CACNB2  CACNB2
#> ENSG00000120129 3.666925e-130 2.202355e-126   DUSP1   DUSP1
#> ENSG00000101347 9.583815e-130 4.317030e-126  SAMHD1  SAMHD1
#> ENSG00000189221 1.098955e-123 3.960194e-120    MAOA    MAOA
#> ENSG00000211445 4.618318e-112 1.386881e-108    GPX3    GPX3