Information on a gene

Assembles information, in HTML format, regarding a gene symbol identifier

Usage

geneinfo_to_html(gene_id, res_de = NULL, col_to_use = "SYMBOL")

Arguments

gene_id

Character specifying the gene identifier for which to retrieve information

res_de

An object containing the results of the Differential Expression analysis workflow (e.g. DESeq2, edgeR or limma). Currently, this can be a DESeqResults object created using the DESeq2 framework. If not provided, the experiment-related information is not shown, and only some generic info on the identifier is displayed. The information about the gene is retrieved by matching on the SYMBOL column, which should be provided in res_de.

col_to_use

The column of your res_de object containing the gene symbols. Default is "SYMBOL"

Value

HTML content related to a gene identifier, to be displayed in web applications (or inserted in Rmd documents)

Details

Creates links to the NCBI and the GeneCards databases

Examples

geneinfo_to_html("ACTB")
#> <b>ACTB</b><br/>Link to the NCBI Gene database: <a href="http://www.ncbi.nlm.nih.gov/gene/?term=ACTB[sym]" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">ACTB@NCBI</a><br/>Link to the GeneCards database: <a href = "https://www.genecards.org/cgi-bin/carddisp.pl?gene=ACTB" target = "_blank" class = "btn btn-primary" style = "color: #ffffff; background-color: #0092AC">ACTB@GeneCards</a><br/>Link to the GTEx Portal: <a href = "https://www.gtexportal.org/home/gene/ACTB" target = "_blank" class = "btn btn-primary" style = "color: #ffffff; background-color: #0092AC"><i class="fa fa-dna"></i>ACTB@GTEX</a><br/>Link to the UniProt Portal: <a href = "https://www.uniprot.org/uniprot/?query=ACTB" target = "_blank" class = "btn btn-primary" style = "color: #ffffff; background-color: #0092AC"><i class="fa fa-spinner"></i>ACTB@UNIPROT</a><br/>Link to related articles on Pubmed: <a href="https://pubmed.ncbi.nlm.nih.gov/?term=ACTB" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">ACTB@Pubmed</a><br/>
geneinfo_to_html("Pf4")
#> <b>Pf4</b><br/>Link to the NCBI Gene database: <a href="http://www.ncbi.nlm.nih.gov/gene/?term=Pf4[sym]" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">Pf4@NCBI</a><br/>Link to the GeneCards database: <a href = "https://www.genecards.org/cgi-bin/carddisp.pl?gene=Pf4" target = "_blank" class = "btn btn-primary" style = "color: #ffffff; background-color: #0092AC">Pf4@GeneCards</a><br/>Link to the GTEx Portal: <a href = "https://www.gtexportal.org/home/gene/Pf4" target = "_blank" class = "btn btn-primary" style = "color: #ffffff; background-color: #0092AC"><i class="fa fa-dna"></i>Pf4@GTEX</a><br/>Link to the UniProt Portal: <a href = "https://www.uniprot.org/uniprot/?query=Pf4" target = "_blank" class = "btn btn-primary" style = "color: #ffffff; background-color: #0092AC"><i class="fa fa-spinner"></i>Pf4@UNIPROT</a><br/>Link to related articles on Pubmed: <a href="https://pubmed.ncbi.nlm.nih.gov/?term=Pf4" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">Pf4@Pubmed</a><br/>