Information on a Gene Ontology identifier

Assembles information, in HTML format, regarding a Gene Ontology identifier

Usage

go_to_html(go_id, res_enrich = NULL)

Arguments

go_id

Character, specifying the GeneOntology identifier for which to retrieve information

res_enrich

A data.frame object, storing the result of the functional enrichment analysis. If not provided, the experiment-related information is not shown, and only some generic info on the identifier is displayed.

Value

HTML content related to a GeneOntology identifier, to be displayed in web applications (or inserted in Rmd documents)

Details

Also creates a link to the AmiGO database

Examples

go_to_html("GO:0002250")
#> <b>GO ID: </b><a href="http://amigo.geneontology.org/amigo/term/GO:0002250" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">GO:0002250@AMIGO</a><br/><b>Pubmed results: </b><a href="https://pubmed.ncbi.nlm.nih.gov/?term=adaptive immune response" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">adaptive immune response@Pubmed</a><br/><b>Term: </b>adaptive immune response<br/><b>Ontology: </b>BP<br/><br/><b>Definition: </b>An immune response mediated by cells expressing specific receptors for antigens produced through a somatic diversification process, and allowing for an enhanced secondary response to subsequent exposures to the same antigen (immunological memory).<br/><b>Synonym: </b>acquired immune response<br/><b>Synonym: </b>immune memory response<br/>
go_to_html("GO:0043368")
#> <b>GO ID: </b><a href="http://amigo.geneontology.org/amigo/term/GO:0043368" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">GO:0043368@AMIGO</a><br/><b>Pubmed results: </b><a href="https://pubmed.ncbi.nlm.nih.gov/?term=positive T cell selection" target="_blank" class="btn btn-primary" style = "color: #ffffff; background-color: #0092AC">positive T cell selection@Pubmed</a><br/><b>Term: </b>positive T cell selection<br/><b>Ontology: </b>BP<br/><br/><b>Definition: </b>The process of sparing immature T cells which react with self-MHC protein complexes with low affinity levels from apoptotic death.<br/><b>Synonym: </b>positive T lymphocyte selection<br/><b>Synonym: </b>positive T-cell selection<br/><b>Synonym: </b>positive T-lymphocyte selection<br/>