mosdef: mostly useful differential expression functions

mostly useful differential expression functions

Details

This package provides functionality to run a number of tasks in the differential expression analysis workflow. This encompasses the most widely used steps, from running various enrichment analysis tools with a unified interface to creating plots and beautifying table components linking to external websites and databases. This streamlines the generation of comprehensive analysis reports.

See also

Author

Maintainer: Federico Marini marinif@uni-mainz.de (ORCID)

Authors:

Leon Dammer lc.dammer@gmail.com (ORCID)