A function checking if your dds contains everything you need

A function checking if your dds contains everything you need

Usage

mosdef_dds_check(dds, verbose = FALSE)

Arguments

dds

A DESeqDataset object created using DESeq2

verbose

Logical, whether to add messages telling the user which steps were taken

Value

An invisible NULL after performing the checks

Examples

library("macrophage")
library("DESeq2")
data(gse, package = "macrophage")

dds_macrophage <- DESeqDataSet(gse, design = ~ line + condition)
#> using counts and average transcript lengths from tximeta
rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)
keep <- rowSums(counts(dds_macrophage) >= 10) >= 6
dds_macrophage <- dds_macrophage[keep, ]
dds_macrophage <- DESeq(dds_macrophage)
#> estimating size factors
#> using 'avgTxLength' from assays(dds), correcting for library size
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing

mosdef_dds_check(dds_macrophage)