A function checking if your res_de contains everything you need

A function checking if your res_de contains everything you need

Usage

mosdef_res_check(res_de, verbose = FALSE)

Arguments

res_de

An object containing the results of the Differential Expression analysis workflow (e.g. DESeq2, edgeR or limma). Currently, this can be a DESeqResults object created using the DESeq2 framework.

verbose

Logical, whether to add messages telling the user which steps were taken

Value

An invisible NULL after performing the checks

Examples

data(res_de_macrophage, package = "mosdef")

mosdef_res_check(res_macrophage_IFNg_vs_naive)
#> Warning: Could not find a 'symbol' column. This can be created for you in the other functions