Create a button for gene symbols in an Rmd

A function to turn Gene Symbols into buttons in an Rmd linking to various Portals for further info about these genes. Current supported portals are: GeneCards, NCBI, GTEx, Uniprot, dbPTM, Human Protein Atlas

Usage

buttonifier(
  df,
  new_cols = c("GC", "UNIPROT"),
  col_to_use = "SYMBOL",
  output_format = "DT"
)

Arguments

df

A dataframe with at least on column with gene Symbols named: SYMBOL

new_cols

At least one of: "GC", "NCBI", "GTEX", "UNIPROT", "dbPTM", "HPA" "PUBMED"

col_to_use

name of the columns were the gene symbols are stored. Default is SYMBOL

output_format

a parameter deciding which output format to return, either a DT:datatable (recommended) or a simple dataframe (DF).In the latter case it is important that if the data is visualized with the datatable function the parameter escape must be set to FALSE

Value

A dataframe or a DT datatable object with columns adding HTML objects that link to websites with further information on the genes in question.

Examples

data(res_airway, package = "mosdef")
res_de <- res_airway
res_df <- deseqresult2df(res_de)
# Subsetting for quicker run
res_df <- res_df[1:100,]
buttonifier(res_df)